requirements.txt
numpy
pyzbar
pillow
